добавил тестирование модуля  shapes
