+++ hw06_testing/comparator/main_test.go
@@ -1,0 +1,18 @@
+package comparator
+
+import "testing"
+
+func Test(t *testing.T) {
+	testCases := []struct {
+		desc string
+	}{
+		{
+			desc: "",
+		},
+	}
+	for _, tC := range testCases {
+		t.Run(tC.desc, func(t *testing.T) {
+
+		})
+	}
+}
